backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
mysqlclient==2.2.7
django-cors-headers==4.9.0